Set GIT_PAGER to avoid requiring manual interference during import
Here's a small patch that helps on a few repos I've converted.
If the output of git-log is too long for one's terminal, git
sends the output to less(1) (or perhaps more(1)) which requires
manual interaction in order to continue with the import.
This patch causes git to use 'cat' instead, which fixed my problems.