Add borings to the global darcs boring file
I had an issue with darcs repositories which would have a versioned
boring file, specified with darcs setpref. This file didn't exclude git,
so when darcs-to-git processed that commit, the working directory would
no longer be clean (as it would stop ignoring the .git directory). This
patch writes those ignores to the global ignore file, which is always
consulted.

This patch only adds the patterns that don't already exist, but doesn't
restore the old state on completion.